requirements.txt
azure-cosmos==4.7.0
email_validator==2.2.0
fastapi==0.115.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
uvicorn==0.32.0
pydantic[email]
